Mostly EntityClass changes
Made it so that the temporary vector which will be pushed into our board vector is defined within the first for loop. without doing such, the rows and columns when the draw board method are disfragmented.

Changed the ints to unsigned so that when comparing them with the sizes of the vectors we get rid of random compiler warnings.

changed the inner loop of the drawboard function so that it doesn't print out indexes of one of the given inner vectors and instead outputs the full amount of "~"s in the board.

Changed the water from "0" to "~".

There is an elusive compiler assertion error when simply running the drawboard method. i need to check whether after these changes this is sitll an issue.