GuessShip and destroyShip has been setup
How i plan to make these two methods work is that guessShip will prompt two numbers from the AI/Player which will act as a set of coordinates. the destroy ship method will then be called on the opposing enemy, checking if these coordinates have a ship. if they do, destroy the ship. if not, do nothing.