Changed coordinate input/validation
I've condensed the check for coordinates that are actually inside the board into a member function of the player class.

How it works is that when calling, you pass in a string related to which coordinate you are looking for(if you're looking for a row number, type "row", and if you're looking for a column number, type "column"), it then asks for an integer and checks whether it exists on the board, looping input if it doesn't.