Fixed some errors and added some new code
fixed the bug where the player could enter a column which didn't exist on the board, causing an exception to be thrown when it was referenced in the board vector.

Added some code to the main function which sets up the player and enemy attributes(although i am going to move all of this into another function at some point).

Tomorrow, i'm going to make the input more user friendly, so that, for example when getting coordinates, you can input 1 instead of 0.